Move example to an example directory
