+++ .npmignore
@@ -1,1 +1,2 @@
 src/
+example/
