+++ example/main.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Sparkline from './d3-react-sparkline'
+import Sparkline from '../src/d3-react-sparkline'
 
 const data = [85, 66, 71, 10, 5, 16, 71, 1, 16, 24, 54, 85, 37, 36, 43, 67, 63, 23, 96, 53, 25]
