+++ example/server.js
@@ -1,8 +1,9 @@
 var webpack = require('webpack')
 var WebpackDevServer = require('webpack-dev-server')
-var config = require('./webpack.config')
+var config = require('../webpack.config')
 
 new WebpackDevServer(webpack(config), {
+  contentBase: __dirname,
   publicPath: config.output.publicPath,
   hot: true,
   historyApiFallback: true
